+++ testData/org/elixir_lang/heex/documentation/component/page_live.ex
@@ -1,0 +1,6 @@
+defmodule MyAppWeb.PageLive do
+  @doc "Renders a clickable button."
+  def button(assigns) do
+    ~H""
+  end
+end