+++ testData/org/elixir_lang/heex/documentation/remote/core_components.ex
@@ -1,0 +1,6 @@
+defmodule MyAppWeb.CoreComponents do
+  @doc "Renders a clickable button."
+  def button(assigns) do
+    ~H""
+  end
+end